toolkit/components/passwordmgr/test/mochitest/test_recipe_login_fields.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for recipes overriding login fields</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script type="application/javascript;version=1.8">
SimpleTest.waitForExplicitFinish();

const PWMGR_COMMON_URL = SimpleTest.getTestFileURL("pwmgr_common.js");
let pwmgrCommonScript = SpecialPowers.loadChromeScript(PWMGR_COMMON_URL);

// Tell the parent to setup test logins.
pwmgrCommonScript.sendAsyncMessage("setupParent", { selfFilling: true });

// When the setup is done, load a recipe for this test.
pwmgrCommonScript.addMessageListener("doneSetup", function doneSetup() {
  pwmgrCommonScript.sendAsyncMessage("loadRecipes", {
    siteRecipes: [{
      hosts: ["mochi.test:8888"],
      usernameSelector: "input[name='uname1']",
      passwordSelector: "input[name='pword2']",
    }],
  });
});

pwmgrCommonScript.addMessageListener("loadedRecipes", function loadedRecipes() {
  // Append the form dynamically so autofill is triggered after setup above.
  document.getElementById("content").innerHTML += `
<!-- form with recipe for the username and password -->
  <form id="form1">
    <input type="text"     name="uname1" oninput="reportFill(this, true)">
    <input type="text"     name="uname2" oninput="reportFill(this, false)">
    <input type="password" name="pword1" oninput="reportFill(this, false)">
    <input type="password" name="pword2" oninput="reportFill(this, true)">
  </form>`;
});

const EXPECTED_FILLS = 4;
let fillCount = 0;

function reportFill(element, expected) {
  ok(expected, `${element.name} was filled`);
  if (++fillCount == EXPECTED_FILLS) {
    pwmgrCommonScript.destroy();
    SimpleTest.finish();
  } else if (fillCount == 2) {
    document.getElementById("content").innerHTML = `
  <!-- Fallback to the default heuristics since the selectors don't match -->
  <form id="form2">
    <input type="text"     name="uname3" oninput="reportFill(this, false)">
    <input type="text"     name="uname4" oninput="reportFill(this, true)">
    <input type="password" name="pword3" oninput="reportFill(this, true)">
    <input type="password" name="pword4" oninput="reportFill(this, false)">
  </form>`;
  } else if (fillCount > EXPECTED_FILLS) {
    ok(false, "Too many fills");
  }
}
</script>

<p id="display"></p>

<div id="content">
  // Forms are inserted dynamically
</div>
<pre id="test"></pre>
</body>
</html>
